Update for Debian buster
The included configuration files are for Debian stretch. These do not
work for buster.

The preseed-xx_XX files were removed, because there is no need for
these. One should simply copy example-preseed.cfg and edit this as
needed. The README was updated accordingly.

The Makefile was updated to create a virtual disk image for QEMU,
otherwise the installation will complain that there is no harddisk.

The Makefile was also updated so that build steps are executed only when
necessary.

I renamed config.txt to Makevars, because that's more common.